+++ src/otx/algo/instance_segmentation/utils/structures/mask/mask_target.py
@@ -8,6 +8,8 @@ Reference : https://github.com/open-mmlab/mmdetection/blob/v3.2.0/mmdet/structur
 
 from __future__ import annotations
 
+import warnings
+
 import numpy as np
 import torch
 from datumaro.components.annotation import Polygon
@@ -62,16 +64,20 @@ def mask_target_single(
     meta_info: dict,
 ) -> Tensor:
     """Compute mask target for each positive proposal in the image."""
+    mask_size = _pair(cfg["mask_size"])
+    if len(gt_masks) == 0:
+        warnings.warn("No ground truth masks are provided!", stacklevel=2)
+        return pos_proposals.new_zeros((0, *mask_size))
+
     if isinstance(gt_masks[0], Polygon):
         crop_and_resize = crop_and_resize_polygons
     elif isinstance(gt_masks, tv_tensors.Mask):
         crop_and_resize = crop_and_resize_masks
     else:
-        msg = f"Unsupported type of masks: {type(gt_masks[0])}"
-        raise NotImplementedError(msg)
+        warnings.warn("Unsupported ground truth mask type!", stacklevel=2)
+        return pos_proposals.new_zeros((0, *mask_size))
 
     device = pos_proposals.device
-    mask_size = _pair(cfg["mask_size"])
     num_pos = pos_proposals.size(0)
     if num_pos > 0:
         proposals_np = pos_proposals.cpu().numpy()